group_commands
added GroupChannel to settings.
this is the channel we use for group commands
like 'dgae' or 'bcca'  we will append a / as a prefix and
if its an EQBC command we append " /" to the end of it otherwise for dannet  leave it as is.

added the drop down to lootnscoot menu for group commands. so you can now send you entire group to sell, tribute, bank, or issue a cleanup on bags from the driver.