SQL update
Migrate loot.ini to a sql table.

Will use the sql table to keep track of global and normal loot rules. buy items are still stored in the char_settings file.

initial creation of the DB does not import the loot.ini file. this is to be done on one character with /lootutils import

All changes and new items will also be sent over actors to the other characters running on this PC so they stay current in their working tables.