CanUse flag, themes on all windows
setting a theme from the console config window will apply to all windows.

Added loot Rule CanUse
This is the default setting for new items that are nodrop we set this before even processing the decision.

this will prevent leaving items we can use but someone else had checked it first and marked it ignore.