performance, and fixes
new items are default set to 'Sell' if they have a sell value

We use our settings to make the decision to loot or not. so an item marked sell that sells for less than our settings we will ignore.

same for items with no sell value and a tribute value.

Major improvements on loading everything on initial loadup.

Added ignorebagslot setting
setting this will ignore anything in that bag slot when it comes to selling\tributeing\banking\destroying