bank changes and some fixes
the bank tradeskills flag will no longer auto flag items to bank

Instead it will be checked when issuing the bank command.

When banking:
check rules and if its 'Bank' toss it into the bank
if the rule is not Global or Personal then check if the item is a TS item, and if we have Bank TS enabled.
if both are true then bank it.

this way we don't have rules flip flopping from this setting. and it still serves a purpose for items without overrides in global or personal rules tables.

This still allows you to manually set items to Bank and it will follow that rule.