feat: command line (#44)
refactored some command line work.

`/lns [normalitem|globalitem|personalitem] [rule] [itemName|qty] [qty]`

Add an item rule to the Items database.
The item will be added to the proper database table based on which is issued normalitem, personalitem, or globalitem.
Qty is only applied if the rule is quest. if an item is on the cursor you can omitt the item name and it will use the item on the cursor.
If the item is already in the database, it will update the quantity