bug: loot decision making work
bug squashing the decision making routine.

Some rework on decision making especially for nodrop and new rules.

I've been chasing occasional looting items when we shouldn't have and narrowed it down to outside scripts / automation as the main cause.

there are plenty of debugging messages and I have yet to see it to enter the loot item when not supposed to anymore.

I have tested with pausing lns until fighting is over and then pausing the automation and unpausing lns. (no issues)
as well as creating a pile of corpses then stopping automation and starting lns. also no issues.