ref: refactor decision making (#49)
after thinking about this the entire 200mile drive there and 2 days of camping...

complete redo of the getRule function

cleaner, easier to work through. Now with Less Spaghetti!

no drop items will not create an initial rule if `lootNoDrop` and `lootNoDropNew` are NOT both enabled. this will keep it as NEW for all chars until the rule is confirmed.